+++ src/jest/browserStorage/index.js
@@ -1,60 +1,62 @@
 export const fakeStorageInstanceFactory = () => (function () {
-    let __keys = [];
-    let __map = {};
+  let __keys = []
+  let __map = {}
 
-    const storageFake = new Proxy(Object.freeze({
+  const storageFake = new Proxy(
+    Object.freeze({
       setItem (key, value) {
-        if (typeof key !== "string") {
-          return;
+        if (typeof key !== 'string') {
+          return
         }
 
-        if (typeof value !== "string") {
-          return;
+        if (typeof value !== 'string') {
+          return
         }
 
-        __keys.push(key);
-        __map[key] = value;
+        __keys.push(key)
+        __map[key] = value
       },
       removeItem (key) {
-        const index = __keys.indexOf(key);
+        const index = __keys.indexOf(key)
         if (index === -1) {
-          return;
+          return
         }
-        __keys.splice(index, 1);
-        delete __map[key];
+        __keys.splice(index, 1)
+        delete __map[key]
       },
       key (keyIndex) {
-        if (typeof keyIndex !== "number") {
-          return null;
+        if (typeof keyIndex !== 'number') {
+          return null
         }
-        return __keys[keyIndex] || null;
+        return __keys[keyIndex] || null
       },
       length: -1,
       clear () {
-        __keys = [];
-        __map = {};
+        __keys = []
+        __map = {}
       },
       getItem (key) {
-        if (typeof key !== "string") {
-          return null;
+        if (typeof key !== 'string') {
+          return null
         }
-        return __map[key] || null;
+        return __map[key] || null
       }
     }),
     {
       get: (target, property) => {
-        if (typeof target[property] !== "number") {
-          return target[property];
+        if (typeof target[property] !== 'number') {
+          return target[property]
         } else {
-          if (property === "length") {
-            return __keys.length;
+          if (property === 'length') {
+            return __keys.length
           }
         }
       }
-    });
+    }
+  )
 
-    storageFake[Symbol.toStringTag] = "Storage";
-    storageFake["constructor"] = 'function Storage() { [native code] }';
+  storageFake[Symbol.toStringTag] = 'Storage'
+  storageFake.constructor = 'function Storage() { [native code] }'
 
-    return storageFake;
-  }());
+  return storageFake
+}())