+++ src/jest/react-transition-group/index.js
@@ -1,7 +1,7 @@
 export const fakeReactJSTransitionGroupFactory = () => {
-    return () => {
-        return {
-            CSSTransition: jest.fn(({ children, in: show }) => (show ? children : null)),
-        }
+  return () => {
+    return {
+      CSSTransition: jest.fn(({ children, in: show }) => (show ? children : null))
     }
-};
+  }
+}