+++ CHANGELOG.md
@@ -1,8 +1,16 @@
-<a name="0.1.0"></a>
-# 0.1.0 (2022-03-02)
+<a name="0.0.1"></a>
+# 0.0.1 (2022-04-16)
 
 ### Feature Added
-- None
+- Added `provisionMockedDateForTests()` helper function.
+- Added `provisionFakeBrowserSessionStorageForTests()` helper function.
+- Added `provisionFakeBrowserLocalStorageForTests()` helper function.
+- Added `provisionFakeBrowserIntersectionObserverForTests()` helper function.
+- Added `provisionMockedNextJSRouterForTests()` helper function.
+- Added `provisionMockedWinstonLoggerForTests()` helper function.
+- Added `provisionMockedPinoLoggerForTests()` helper function.
+- Added `provisionMockedReactTransitionGroupForTests()` helper function.
+- Added `provisionFakeDateForTests()` helper function.
 
 ### Bug Fixes
 - None